added token verification friendship controller

+++ DrGym/src/main/java/com/drgym/drgym/controller/FriendshipController.java
@@ -1,11 +1,12 @@
 package com.drgym.drgym.controller;
 
-import com.drgym.drgym.model.Friendship;
-import com.drgym.drgym.model.FriendshipInvitation;
+import jakarta.servlet.http.HttpServletRequest;
 import com.drgym.drgym.service.FriendshipService;
 import com.drgym.drgym.service.FriendshipService.UserFriendDTO;
 import com.drgym.drgym.service.FriendshipService.FriendRequestDTO;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
@@ -19,39 +20,70 @@ public class FriendshipController {
     @Autowired
     private FriendshipService friendshipService;
 
-    @GetMapping("friendsinfo/{username}")
-    public Map<String, List<?>> getUserFriendsAndInvitations(@PathVariable String username) {
+    @Autowired
+    private UserController userController;
+
+    @GetMapping("/friendsinfo/{username}")
+    public ResponseEntity<?> getUserFriendsAndInvitations(@PathVariable String username, HttpServletRequest request) {
+        if (!userController.tokenOwner(username, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+
         List<UserFriendDTO> friends = friendshipService.getUserFriends(username);
         List<FriendRequestDTO> invitations = friendshipService.getUserFriendshipInvitations(username);
         Map<String, List<?>> response = new HashMap<>();
         response.put("friends", friends);
         response.put("invitations", invitations);
-        return response;
+        return ResponseEntity.ok(response);
     }
 
     @GetMapping("/isFriend/{user1}/{user2}")
-    public boolean isFriend(@PathVariable String user1, @PathVariable String user2) {
-        return friendshipService.areFriends(user1, user2);
+    public ResponseEntity<?> isFriend(@PathVariable String user1, @PathVariable String user2, HttpServletRequest request) {
+        if (!userController.tokenOwnerOrFriend(user1, request) && !userController.tokenOwnerOrFriend(user2, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+        boolean areFriends = friendshipService.areFriends(user1, user2);
+        return ResponseEntity.ok(areFriends);
     }
 
     @PostMapping("/sendRequest")
-    public String sendFriendRequest(@RequestBody FriendRequestDTO friendRequest) {
-        return friendshipService.sendFriendRequest(friendRequest.getSender(), friendRequest.getReceiver());
+    public ResponseEntity<String> sendFriendRequest(@RequestBody FriendRequestDTO friendRequest, HttpServletRequest request) {
+        if (!userController.tokenOwner(friendRequest.getSender(), request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+        String result = friendshipService.sendFriendRequest(friendRequest.getSender(), friendRequest.getReceiver());
+        return ResponseEntity.ok(result);
     }
 
     @PostMapping("/acceptRequest")
-    public String acceptFriendRequest(@RequestParam Long invitationId) {
-        return friendshipService.acceptFriendRequest(invitationId);
+    public ResponseEntity<String> acceptFriendRequest(@RequestParam Long invitationId, HttpServletRequest request) {
+        String username = friendshipService.getInvitationReceiver(invitationId);
+        if (!userController.tokenOwner(username, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+        String result = friendshipService.acceptFriendRequest(invitationId);
+        return ResponseEntity.ok(result);
     }
 
     @PostMapping("/rejectRequest")
-    public String rejectFriendRequest(@RequestParam Long invitationId) {
-        return friendshipService.rejectFriendRequest(invitationId);
+    public ResponseEntity<String> rejectFriendRequest(@RequestParam Long invitationId, HttpServletRequest request) {
+        String username = friendshipService.getInvitationReceiver(invitationId);
+        if (!userController.tokenOwner(username, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+        String result = friendshipService.rejectFriendRequest(invitationId);
+        return ResponseEntity.ok(result);
     }
 
     @DeleteMapping("/removeFriend")
-    public String removeFriend(@RequestBody RemoveFriendRequest removeFriendRequest) {
-        return friendshipService.removeFriend(removeFriendRequest.getUser1(), removeFriendRequest.getUser2());
+    public ResponseEntity<String> removeFriend(@RequestBody RemoveFriendRequest removeFriendRequest, HttpServletRequest request) {
+        String user1 = removeFriendRequest.getUser1();
+        String user2 = removeFriendRequest.getUser2();
+        if (!userController.tokenOwnerOrFriend(user1, request) && !userController.tokenOwnerOrFriend(user2, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+        String result = friendshipService.removeFriend(removeFriendRequest.getUser1(), removeFriendRequest.getUser2());
+        return ResponseEntity.ok(result);
     }
 
     public static class RemoveFriendRequest {
